requirements.txt
absl-py==0.9.0
astunparse==1.6.3
cachetools==4.1.0
certifi==2020.4.5.1
chardet==3.0.4
click==7.1.2
cmake==3.17.2
configparser==5.0.0
coverage==5.1
cycler==0.10.0
dlib==19.19.0
gast==0.3.3
google-auth==1.15.0
google-auth-oauthlib==0.4.1
google-pasta==0.2.0
grpcio==1.29.0
h5py==2.10.0
idna==2.9
imutils==0.5.3
joblib==0.15.1
Keras==2.3.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
keyboard==0.13.5
kiwisolver==1.2.0
Markdown==3.2.2
matplotlib==3.2.1
nose2==0.9.2
numpy==1.18.4
oauthlib==3.1.0
opencv-python==4.2.0.34
opt-einsum==3.2.1
pandas==1.0.3
Pillow==7.1.2
protobuf==3.12.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2020.1
PyYAML==5.3.1
requests==2.23.0
requests-oauthlib==1.3.0
rsa==4.0
scikit-learn==0.23.1
scipy==1.4.1
six==1.15.0
sklearn==0.0
tensorboard==2.2.1
tensorboard-plugin-wit==1.6.0.post3
tensorflow==2.2.0
tensorflow-estimator==2.2.0
termcolor==1.1.0
threadpoolctl==2.0.0
urllib3==1.25.9
Werkzeug==1.0.1
wrapt==1.12.1
